requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
sqlalchemy==2.0.48
asyncpg==0.30.0
pydantic==2.12.5
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
